strata_agent/pyproject.toml
[project]
name = "strata-agent"
version = "0.1.0"
description = "Strata SMB Connector Agent"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "vulture>=2.13",
    "bandit[toml]>=1.7.0",
    "types-PyYAML",
]

[project.scripts]
strata-agent = "strata_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Tooling config inherited from root pyproject.toml
